// src/util/enum.ts
export enum Gamemode {
	Survival = 0x00,
	Creative = 0x01,
	Adventure = 0x02,
	Spectator = 0x03,
}

export enum EntityType {
	Allay = 2,
	AreaEffectCloud = 3,
	ArmorStand = 4,
	Arrow = 5,
	Axolotl = 6,
	Bat = 7,
	Bee = 8,
	Blaze = 9,
	BlockDisplay = 10,
	Boat = 11,
	Camel = 12,
	Cat = 13,
	CaveSpider = 14,
	ChestBoat = 15,
	ChestMinecart = 16,
	Chicken = 17,
	Cod = 18,
	CommandBlockMinecart = 19,
	Cow = 20,
	Creeper = 21,
	Dolphin = 22,
	Donkey = 23,
	DragonFireball = 24,
	Drowned = 25,
	Egg = 26,
	ElderGuardian = 27,
	EndCrystal = 28,
	EnderDragon = 29,
	EnderPearl = 30,
	Enderman = 31,
	Endermite = 32,
	Evoker = 33,
	EvokerFangs = 34,
	ExperienceBottle = 35,
	ExperienceOrb = 36,
	EyeOfEnder = 37,
	FallingBlock = 38,
	FireworkRocket = 39,
	Fox = 40,
	Frog = 41,
	FurnaceMinecart = 42,
	Ghast = 43,
	Giant = 44,
	GlowItemFrame = 45,
	GlowSquid = 46,
	Goat = 47,
	Guardian = 48,
	Hoglin = 49,
	HopperMinecart = 50,
	Horse = 51,
	Husk = 52,
	Illusioner = 53,
	Interaction = 54,
	IronGolem = 55,
	Item = 56,
	ItemDisplay = 57,
	ItemFrame = 58,
	Fireball = 59,
	LeashKnot = 60,
	LightningBolt = 61,
	Llama = 62,
	LlamaSpit = 63,
	MagmaCube = 64,
	Marker = 65,
	Minecart = 66,
	Mooshroom = 67,
	Mule = 68,
	Ocelot = 69,
	Painting = 70,
	Panda = 71,
	Parrot = 72,
	Phantom = 73,
	Pig = 74,
	Piglin = 75,
	PiglinBrute = 76,
	Pillager = 77,
	PolarBear = 78,
	Potion = 79,
	Pufferfish = 80,
	Rabbit = 81,
	Ravager = 82,
	Salmon = 83,
	Sheep = 84,
	Shulker = 85,
	ShulkerBullet = 86,
	Silverfish = 87,
	Skeleton = 88,
	SkeletonHorse = 89,
	Slime = 90,
	SmallFireball = 91,
	Sniffer = 92,
	SnowGolem = 93,
	Snowball = 94,
	SpawnerMinecart = 95,
	SpectralArrow = 96,
	Spider = 97,
	Squid = 98,
	Stray = 99,
	Strider = 100,
	Tadpole = 101,
	TextDisplay = 102,
	Tnt = 103,
	TntMinecart = 104,
	TraderLlama = 105,
	Trident = 106,
	TropicalFish = 107,
	Turtle = 108,
	Vex = 109,
	Villager = 110,
	Vindicator = 111,
	WanderingTrader = 112,
	Warden = 113,
	Witch = 114,
	Wither = 115,
	WitherSkeleton = 116,
	WitherSkull = 117,
	Wolf = 118,
	Zoglin = 119,
	Zombie = 120,
	ZombieHorse = 121,
	ZombieVillager = 122,
	ZombifiedPiglin = 123,
	Player = 124,
	FishingBobber = 125,
}

export enum Pose {
	STANDING = 0,
	FALL_FLYING = 1,
	SLEEPING = 2,
	SWIMMING = 3,
	SPIN_ATTACK = 4,
	SNEAKING = 5,
	LONG_JUMPING = 6,
	DYING = 7,
	CROAKING = 8,
	USING_TONGUE = 9,
	SITTING = 10,
	ROARING = 11,
	SNIFFING = 12,
	EMERGING = 13,
	DIGGING = 14,
}

export enum PlayerAction {
	StartSneaking = 0,
	StopSneaking = 1,
	LeaveBed = 2,
	StartSprinting = 3,
	StopSprinting = 4,
	StartJumpWithHorse = 5,
	StopJumpWithHorse = 6,
	OpenVehicleInventory = 7,
	StartFlyingWithElytra = 8,
}

export enum MetadataType {
	Byte = 0,
	VarInt = 1,
	VarLong = 2,
	Float = 3,
	String = 4,
	TextComponent = 5,
	OptionalTextComponent = 6,
	Slot = 7,
	Boolean = 8,
	Rotations = 9,
	Position = 10,
	OptionalPosition = 11,
	Direction = 12,
	OptionalUUID = 13,
	BlockState = 14,
	OptionalBlockState = 15,
	NBT = 16,
	Particle = 17,
	VillagerData = 18,
	OptionalVarInt = 19,
	Pose = 20,
	CatVariant = 21,
	FrogVariant = 22,
	OptionalGlobalPosition = 23,
	PaintingVariant = 24,
	SnifferState = 25,
	Vector3 = 26,
	Quaternion = 27,
}

export enum EntityAnimation {
	SwingMainArm = 0,
	LeaveBed = 2,
	SwingOffhand = 3,
	CriticalEffect = 4,
	MagicCriticalEffect = 5,
}